requirements.txt
openai
python-dotenv
imapclient
email-validator 